#316 Get cache folder functionality was moved to startup
